Fixed more tests. Moved form tests to where existing test was.
