+++ tests/frontend/test_forms.py
@@ -1,6 +1,5 @@
 # Standard library
 from unittest import mock
-from unittest.mock import patch
 
 # Third-party
 import pytest
@@ -416,131 +415,3 @@ def test_clean_namespace(env):
     form.cleaned_data = {"namespace": f"my-namespace-{env}"}
 
     assert form.clean_namespace() == "my-namespace"
-
-
-@pytest.mark.django_db
-def test_ip_allowlist_form_missing_name():
-    data = {
-        "name": "",
-        "allowed_ip_ranges": "192.168.0.0/28",
-    }
-    f = forms.IPAllowlistForm(data)
-    assert f.errors["name"] == ["This field is required."]
-
-
-@patch.object(aws.AWSIdentityStore, "get_user_id")
-@pytest.mark.parametrize(
-    "user, data, expected_groups, expected_result",
-    [
-        (
-            "superuser",
-            {"enable_quicksight": ["quicksight_compute_reader"]},
-            ["quicksight_compute_author", "azure_holding"],
-            2,
-        ),
-        (
-            "normal_user",
-            {"enable_quicksight": ["quicksight_compute_reader"]},
-            ["quicksight_compute_reader", "azure_holding"],
-            2,
-        ),
-        (
-            "superuser",
-            {"enable_quicksight": ["quicksight_compute_author"]},
-            ["quicksight_compute_author", "azure_holding"],
-            2,
-        ),
-        (
-            "normal_user",
-            {"enable_quicksight": ["quicksight_compute_author"]},
-            ["quicksight_compute_author", "azure_holding"],
-            2,
-        ),
-    ],
-)
-def test_quicksight_form_add_to_groups(
-    get_user_id,
-    identity_store_user_setup,
-    users,
-    identity_store,
-    identity_store_id,
-    group_ids,
-    user,
-    data,
-    expected_groups,
-    expected_result,
-):
-    """
-    Tests adding a user to the correct group plus the azure holding group.
-    Super users should not be added to these groups
-    """
-
-    test_user = users[user]
-    get_user_id.return_value = test_user.identity_center_id
-
-    form = forms.QuicksightAccessForm(user=test_user)
-    form.cleaned_data = data
-
-    form.grant_access()
-    response = identity_store.list_group_memberships_for_member(
-        IdentityStoreId=identity_store_id,
-        MemberId={"UserId": test_user.identity_center_id},
-    )
-
-    assert len(response["GroupMemberships"]) == expected_result
-
-    groups = [group_ids[group] for group in expected_groups]
-
-    for group_membership in response["GroupMemberships"]:
-        if group_membership["GroupId"] not in groups:
-            raise AssertionError
-
-
-@patch.object(aws.AWSIdentityStore, "get_group_membership_id")
-@patch.object(aws.AWSIdentityStore, "get_user_id")
-@pytest.mark.parametrize(
-    "user, data, expected_result",
-    [
-        (
-            "quicksight_compute_author",
-            {"enable_quicksight": [""]},
-            1,
-        ),
-        (
-            "quicksight_compute_reader",
-            {"enable_quicksight": [""]},
-            1,
-        ),
-    ],
-)
-def test_quicksight_form_remove_from_group(
-    get_user_id,
-    get_group_membership_id,
-    identity_store_user_setup,
-    users,
-    identity_store,
-    identity_store_id,
-    group_ids,
-    user,
-    data,
-    expected_result,
-):
-    """
-    Tests removing a user from a group. Should still be part of azure holding group
-    """
-
-    test_user = users[user]
-    get_user_id.return_value = test_user.identity_center_id
-    get_group_membership_id.return_value = test_user.group_membership_id
-
-    form = forms.QuicksightAccessForm(user=test_user)
-    form.cleaned_data = data
-
-    form.grant_access()
-    response = identity_store.list_group_memberships_for_member(
-        IdentityStoreId=identity_store_id,
-        MemberId={"UserId": test_user.identity_center_id},
-    )
-
-    assert len(response["GroupMemberships"]) == expected_result
-    assert response["GroupMemberships"][0]["GroupId"] == group_ids["azure_holding"]
